Converted FileDrop into a directive.  Created FileSelector component to work with the FileDropDirective.  Reimplemented FileDrop component using new directive and FileSelector component.
