+++ src/lib/ngx-file-selector/file-drop-selector.component.scss
@@ -1,0 +1,3 @@
+.ngx-file-drop__file-input {
+  display: none;
+}
